requirements.txt
discord.py~=1.3.4
tinydb
tzlocal
pytz